Refactors and supporting charts & dashboards
