+++ metadata-dao-impl/restli-dao/src/main/java/com/linkedin/metadata/dao/TagActionRequestBuilders.java
@@ -1,4 +1,4 @@
-package com.linkedin.metadata.dao.internal;
+package com.linkedin.metadata.dao;
 
 import com.linkedin.common.urn.TagUrn;
 import com.linkedin.metadata.dao.BaseActionRequestBuilder;
